refactor: move orphan detection logic from CLI to trash/xdg package
OrphanedFile type, FindOrphanedTrashInfoFiles(), and
ParseTrashInfoFile() are XDG trash domain logic that was incorrectly
placed in the CLI package. Move them to trash/xdg/orphan.go with
their tests in orphan_test.go. CLI now delegates to xdg package.
