+++ internal/cli/prune.go
@@ -5,8 +5,6 @@ import (
 	"fmt"
 	"log/slog"
 	"os"
-	"path/filepath"
-	"strings"
 	"time"
 
 	"github.com/fatih/color"
@@ -30,18 +28,6 @@ var (
 	ErrInvalidDurationNum  = errors.New("duration must be a positive number")
 	ErrInvalidDurationUnit = errors.New("unsupported duration unit")
 )
-
-// OrphanedFile represents an orphaned metadata file with additional details
-type OrphanedFile struct {
-	Path          string
-	DeletedAt     time.Time
-	OriginalPath  string
-	TrashInfoPath string
-	TrashInfoName string
-}
-
-func (o OrphanedFile) GetName() string         { return o.TrashInfoName }
-func (o OrphanedFile) GetDeletedAt() time.Time { return o.DeletedAt }
 
 // Prune handles the pruning of trash contents
 // It processes multiple subcommands for cleaning up the trash
@@ -75,49 +61,6 @@ func (c *CLI) Prune(args []string) error {
 		durations = append(durations, duration)
 	}
 	return c.permanentlyDeleteByTimeRange(durations)
-}
-
-// findOrphanedTrashInfoFiles finds .trashinfo files without corresponding files
-// Returns a list of paths to orphaned .trashinfo files
-func findOrphanedTrashInfoFiles(trashDir string) ([]OrphanedFile, error) {
-	infoDir := filepath.Join(trashDir, "info")
-	filesDir := filepath.Join(trashDir, "files")
-
-	var orphanedFiles []OrphanedFile
-
-	entries, err := os.ReadDir(infoDir)
-	if err != nil {
-		return nil, fmt.Errorf("failed to read info directory: %w", err)
-	}
-
-	for _, entry := range entries {
-		// Skip non-regular files and non-.trashinfo files
-		if !entry.Type().IsRegular() || !strings.HasSuffix(entry.Name(), ".trashinfo") {
-			continue
-		}
-
-		if strings.HasPrefix(entry.Name(), "._") {
-			// exclude mac resource fork
-			slog.Debug("skipped mac resource fork of .trashinfo", "path", entry.Name())
-			continue
-		}
-
-		// Remove .trashinfo suffix to get the corresponding file name
-		fileName := strings.TrimSuffix(entry.Name(), ".trashinfo")
-		trashInfoPath := filepath.Join(infoDir, entry.Name())
-
-		// Check if corresponding file exists in files directory
-		_, err := os.Stat(filepath.Join(filesDir, fileName))
-		if os.IsNotExist(err) {
-			metadata, parseErr := parseTrashInfoFile(trashInfoPath)
-			if parseErr != nil {
-				continue
-			}
-			orphanedFiles = append(orphanedFiles, metadata)
-		}
-	}
-
-	return orphanedFiles, nil
 }
 
 // permanentlyDeleteByTimeRange removes files from trash based on their age.
@@ -240,10 +183,10 @@ func (c *CLI) removeOrphanedMetadata() error {
 		return fmt.Errorf("failed to get trash dirs: %w", err)
 	}
 
-	var orphanedFiles []OrphanedFile
+	var orphanedFiles []xdg.OrphanedFile
 	for _, trashDir := range trashDirs {
 		slog.Debug("pruning orphaned trashinfo", "trashDir", trashDir)
-		files, err := findOrphanedTrashInfoFiles(trashDir)
+		files, err := xdg.FindOrphanedTrashInfoFiles(trashDir)
 		if err != nil {
 			slog.Error("failed to find orphaned metadata in trash dir", "dir", trashDir, "error", err)
 			continue
@@ -290,36 +233,4 @@ func (c *CLI) removeOrphanedMetadata() error {
 
 	fmt.Printf("Successfully removed %d orphaned metadata files.\n", len(orphanedFiles))
 	return nil
-}
-
-// parseTrashInfoFile parses a .trashinfo file and returns an OrphanedFile
-func parseTrashInfoFile(path string) (OrphanedFile, error) {
-	content, err := os.ReadFile(path)
-	if err != nil {
-		return OrphanedFile{}, err
-	}
-
-	lines := strings.Split(string(content), "\n")
-	var deletedAt time.Time
-	var originalPath string
-
-	for _, line := range lines {
-		if after, ok := strings.CutPrefix(line, "Path="); ok {
-			originalPath = after
-		}
-		if after, ok := strings.CutPrefix(line, "DeletionDate="); ok {
-			deletedAtStr := after
-			deletedAt, err = time.Parse("2006-01-02T15:04:05", deletedAtStr)
-			if err != nil {
-				return OrphanedFile{}, err
-			}
-		}
-	}
-
-	return OrphanedFile{
-		DeletedAt:     deletedAt,
-		OriginalPath:  originalPath,
-		TrashInfoPath: path,
-		TrashInfoName: filepath.Base(path),
-	}, nil
 }
